+++ packages/utils/src/json/index.ts
@@ -1,0 +1,16 @@
+import { jsonParse } from "./parse";
+import { jsonSafeParse } from "./safe-parse";
+import { jsonStringify } from "./stringify";
+
+/**
+ * Utility functions for {@link JSON}.
+ *
+ * @since 0.0.9
+ */
+export const Json = {
+	parse: jsonParse,
+	safeParse: jsonSafeParse,
+    stringify: jsonStringify,
+};
+
+export { jsonParse, jsonSafeParse, jsonStringify };